kernel: clippy --fix small-count style lints (14 more lints)
Force-applied via `clippy --fix --broken-code` restricted to a safe
subset; no manual cleanup needed this round.

  vec_init_then_push       ~13 fixes  Vec::new() + .push(...) → vec![]
  redundant_else            ~6        unneeded else after return
  redundant_pattern_matching ~6       matches!(x, Some(_)) → x.is_some()
  option_option              ~6        Option<Option<T>> sugar refinements
  useless_conversion         ~7        .into() removed where types match
  single_char_pattern        ~6        ".starts_with(\"a\")" → ".starts_with('a')"
  while_let_loop             ~6        `while let Some(x) = it.next() { … }` patterns
  manual_strip               ~15       starts_with()+slice → strip_prefix
  manual_clamp               ~8        chained min/max → .clamp(lo, hi)
  op_ref                     ~8        &x op y where ref unnecessary
  no_effect_underscore_binding ~10
  doc_overindented_list_items ~10
  duplicated_attributes      ~9
  unnecessary_literal_bound  ~6

26 files touched.  Build clean.

+++ kernel/src/ipc/channel.rs
@@ -1108,7 +1108,7 @@ pub fn recv_timeout(handle: ChannelHandle, timeout_ns: u64) -> KernelResult<Mess
         }
     }
 
-    let timer_handle = crate::hrtimer::schedule_ns(timeout_ns, timeout_wake, u64::from(task_id));
+    let timer_handle = crate::hrtimer::schedule_ns(timeout_ns, timeout_wake, task_id);
 
     // Block loop: try to receive, block if empty, re-check on wake.
     loop {
